chore(global): :hammer: change debug, rename files
