--- .vscode/launch.json
@@ -1,41 +1,0 @@
-{
-  // Use IntelliSense to learn about possible attributes.
-  // Hover to view descriptions of existing attributes.
-  // For more information, visit: https://go.microsoft.com/fwlink/?linkid=830387
-  "version": "0.2.0",
-  "configurations": [
-    {
-      "name": "Next.js: debug server-side",
-      "type": "node-terminal",
-      "request": "launch",
-      "command": "pnpm dev"
-    },
-    {
-      "name": "Next.js: debug client-side",
-      "request": "launch",
-      "type": "chrome",
-      "runtimeExecutable": "/Applications/Brave Browser.app/Contents/MacOS/Brave Browser",
-      // "runtimeArgs": ["--remote-debugging-port=9222"],
-      "port": 9222,
-      "url": "http://localhost:3000",
-      "webRoot": "${workspaceFolder}",
-      "outFiles": [
-        "${workspaceFolder}/**/*.js",
-        "!**/node_modules/**"
-      ]
-    },
-    {
-      "name": "Next.js: debug full stack",
-      "type": "node-terminal",
-      "request": "launch",
-      "command": "pnpm run dev",
-      "outputCapture": "std",
-      // "console": "integratedTerminal",
-      "serverReadyAction": {
-        "pattern": "started server on .+, url: (https?://.+)",
-        "action": "startDebugging",
-        "name": "Next.js: debug client-side",
-      }
-    }
-  ]
-}
