--- apps/web/pages/_document.js
@@ -1,20 +1,0 @@
-import { createGetInitialProps } from "@joekarow/netwerkr-ui/mantine/next";
-import Document, { Head, Html, Main, NextScript } from 'next/document'
-
-const getInitialProps = createGetInitialProps();
-
-export default class _Document extends Document {
-    static getInitialProps = getInitialProps;
-
-    render () {
-        return (
-            <Html>
-                <Head />
-                <body>
-                    <Main />
-                    <NextScript />
-                </body>
-            </Html>
-        );
-    }
-}
