+++ apps/web/pages/_document.jsx
@@ -1,0 +1,20 @@
+import { createGetInitialProps } from '@joekarow/netwerkr-ui/mantine/next';
+import Document, { Head, Html, Main, NextScript } from 'next/document';
+
+const getInitialProps = createGetInitialProps();
+
+export default class _Document extends Document {
+  static getInitialProps = getInitialProps;
+
+  render() {
+    return (
+      <Html>
+        <Head />
+        <body>
+          <Main />
+          <NextScript />
+        </body>
+      </Html>
+    );
+  }
+}
